Updated to use getSystemResource() as per Edward's suggestion. Since read_short_table is a static method, we can't use getClass, the error is:
JavaParser.java:211: Can't make static reference to method java.lang.Class getClass() in class java.lang.Object.
   getClass().getResourceAsStream(filename + ".tbl");

Instead, we use:
   ClassLoader.getSystemResourceAsStream(
       "ptolemy/lang/java/" + filename + ".tbl");

If we read in the .tbl files, then we try to write out
the .bin files using the old method that uses RandomAccessFile

In the perfect world, we would like to write to a URL,
but instead we just write to the current directory.
Note that this will cause problems if we are running this
as an applet.  I tried writing to a URL as per
http://www.javasoft.com/docs/books/tutorial/networking/urls/readingWriting.html
but got
java.net.UnknownServiceException: protocol doesn't support output