Type-polymorphic IIR from Research In Motion:
We sorely needed a type-polymorphic IIR, specifically for the Complex
domain, so we finally reworked ptolemy.actor.lib.IIR to be type-polymorphic
(before this, we've been using a hacked ComplexIIR...)

We'd like to contribute this back and hope you will find it useful. Now IIR
should work over any algebraic ring of Tokens... :-)

A note about the _denominator[0] override to "one" in attributeChanged()...
This is not really necessary, since _denominator[0] itself is not used in
the computation. A better solution would be to normalize the entire
denominator using _denominator[0] (like matlab does), but that would require
a divide() operation and reduce the range of applications so we left it as
is...

Regards from our DSP Group that includes Aleksandar Necakov who did and
tested this change...