Three preprocessors are added.
globalVariable_.. is used to model the global variables into the inputs or outputs
of every hybrid automaton;

localVarialbe_.. is used to model the some of the local variables into parameters.
the reason is to keep the states consistent; some local variables are modeled into
outputs.

slim_prepro.. is used to get rid of redundant nodes generated by the former two
preprocessors.

Also, the shsif.xml is the original one from Gabor, which will be used as the demo.

Major changes are made to hsif.xsl.