Changes... Moved CActorBase, CPort and CActor.xml and makefile from ptolemy.apps.cgc to ptolemy.domains.giotto.cgc...
Reintroduced the second initializer in CActor because the moml parser
was calling the second one.

In GiottoCEmachineFrameworkGenerator -
major changes
Changed the structure to become a subclass of GiottoCodeGenerator
since it needed the giotto code also to be generated.
Overrides the _getTypeString method to return the actual type
and not token_port. As a result of having to do this, makes
all static methods instance methods so that they can call the other
instance methods which had to be made instanc because they had to be
made overridable