added time resolution as a private variable to the Director class. therefore, different actors in the same model use the same time resolution.
improved the method comment for setCurrentTime to reflect whether the new time 
can be eariler than the current time. By this way, the API is more meanlingful.

although the currentTime variable in Director class is protected, it is
not a good idea to for subclasses to modify it directly with assignment statement. 
A good way for modification is to override the setCurrentTime method in
the base Director class and use the new setCurrentTime method. 

Ideally, the currentTime should be a private variable and the base class
does not constrain the possible assignments in the setCurrentTime method. 
Subclasses override the method to add constraints...