Added a concatenate(ArrayToken, ArrayToken) function.
This function calls ArrayToken.arrayValue() for each of its arguments,
creating two intermediate copies of the array which are not necessary.
Might want to dispatch to (for instance) ArrayToken.concatenate(ArrayToken)
to avoid this inefficiency.

It would be nice to be able to concatenate an arbitrary number of arrays.
Is there a mechanism for having variable length argument lists of built-in
functions?  If not, one approach would be to have a single-argument version
that would behave like concatenate({{1,2,3},{4,5}}) --> {1,2,3,4,5}.