Added Andreas Traupe <Andreas.Traupe@orat.de> renameLegend() "I am missing a method that doesn't add a new legend but renames an existing one. Remove followed by add isn't good for me because it changes the order in the legend." _clear(): call Vector.clear() instead of throwing away the old Vector: "This method is implemented by throwing away a Vector object and create a new empty one. Instead the old Vector object could be cleared.
Code comment states: "Vector.clear() is new in JDK1.2, so we use just
create a new Vector here so that we can compile this with JDK1.1 for use
in JDK1.1 browsers".

However install.htm states: "Ptplot 5.3 requires Java 1.2 or later."

AND: while Vector.clear() is a JDK1.2 method Vector.removeAllElements()
is a JDK1.0 method.

(This is nothing important. I just happened to stumble over it.)