In the previous implemention of the fire() method, an integrator sends outputs (_tentativeState, which is calculated from the previous firing)  if the output port (far receivers) has an unknown status. This constraint is kind of loose and problematic.
In particular, if a composite actor (which contains one or more integrators) appears multiple times in a schedule, each time the composite actor is fired, all the inside receivers are reset, then the integrators will produce the newly calculated outputs at a wrong time (the outputs are scheduled to be produced at a later time).

A few local variables _localRound and _localRoundOutput are added such that the communication between the directors and integrators are more tight. In particular, the integrators will synchronize its local round count to the solver's round count.