The local variable _actorsFired cannot be cleared in the postfire() method because the derived class HybridModalDirector relies on this list for suggesting step size. We clear the list in the prefire method instead.
Note that in ContinuousDirector, actors are postfired before the next suggested step size is queried, because the postfire method may invok fireAt method, which essentially creates breakpoints (and then inevitably affect the choice of next suggested step size).

Fixed a bug in the postfire method where if any refinement returns false from the postfire method, this director returns false in the postfire method. This is tested in the stopTime and stopTime2 under the directory ptII\ptolemy\domains\continuous\test\auto.