Refactored Tokens to add a ConceptToken and PartiallyOrderedToken.
This means that now any token that implements the PartiallyOrderedToken
interface can get the inequality comparisons (<, <=, >, and >=) in the
Ptolemy expression language for free, rather than being limited to
ScalarTokens as they were previously.

I'm not sure if this functionality would be better strucutred as an
actual Token below ptolemy.data.Token and above ScalarToken and
ConceptToken, rather than the interface that it is now.