Added corrected ifc analysis.
This analysis is similar to the ifc analysis from
"Static Monotonicity Analysis for lambda-definable
Functions over Lattices" (p. 145), but with the
unsoundness bug corrected.  This is still not
enough to prove monotonicity of most of our
dimension system expressions, however, since it
assumes that both branches of a monotonic
conditional are monotonic, which is not true in
the case of, for example:
  (x <= Unknown) ? Unknown :
  (x <= Acceleration) ? Velocity :
  (X <= Velocity) ? Postiion :
  Conflict
My next steps will be to extend the analysis so
that it works in cases such as these.