Removed actorConstraintType.
Now all default constraints (that is, top level
constraints shared by all actors that don't have
specific adapters definted) are determied by the
solver strategy, which is currently drawn from
"forward", "backward", and "bidirectional".

My idea here is that rather than support an
exponential number of combinations from the
start, which will be impossible to test
exhaustively, and likely intimidating to the end
user (heck, even intimidating to me!), we can
just have configurations for use casese that we
know to be useful and add to that as necessary
for supporting new cases as they appear.