Added concept names as constants to parser.
Not sure if this is the best approach, since it
seems like it might cause subtle bugs elsewhere in
Ptolemy if, for example, there were a token or
value somewhere else that happened to have the
same name as one of the concepts.  For now,
however, this is the simplest way to get this
functionality.