Genericized CompositeEntity.entityList.
This method previously took a raw Class and
returned a raw List, but the entire point of the
functionality of the method was to return a list
whose elements were of the type of the given
class. I've now added that information to the
signature of the method through Java generics.

I realize that this is a change to ptolemy.kernel,
which is used by Kepler and other projects, but I
can't imagine any situation where someone would
have used this method in a way other than the
newly specified signature.  Hopefully.