Made CarTrackingSimple a test rather than a demo.
For some reason the CollectLatticeOntologySolverDetails
was causing the test to fail, but I think that even with
it removed, it makes a valid and useful test case.

Also, I had to change a URL to make it absolute
rather than relative so that moving the location
of the moml file didn't mess things up.