Monotonicity analysis over product lattices.
This is only a first step, and only contains a
proof of concept type example.  The main problem
with the approach so far is that the monotonicity
analysis assumes a single domain ontology, whereas
in this case the domain and range of most
functions over product lattices are different.  In
the example checked in here, for example, the
domain should be DomainLattice x DomainLattice
whereas the range should be just DomainLattice.
As far as I can see, the simplest way to support
this is to just make the monotonicity analysis be
able to handle analyses in which the domain and
range ontologies are different.