Added nonLatticeReason to DirectedAcyclicGraph.
This method is a strict generalization of
isLattice that also returns information about why
the graph is not a lattice in the case that the
graph is not a lattice.

Note that the interface for nonLatticeReason() is
by no means finalized! It may make more sense to
have a more machine readable result than the human
readable string that is returned now.