Branch for HandSimDroid team
We plan to create ptServer & ptDroid subsystems for client and server parts respectively.  ptServer will be part of ptII project and will reside in a separate folder.  We will try to do the same for ptDroid but if it does not work due to peculiarities of Android build system we will have to create a separate eclipse project in the handsimdroid branch.