Add a failing test regarding backward inference.
It seems that the current implementation of backward and forward
inference in the ontologies package refers only to inference over an
actor. Relations, however, also can be inferred over in a forwards or
backwards manner. This test case shows an example model where the
current behavior doesn't meet the intuitive defintion of a "backward
analysis" because of the bidirectional behavior over relations.
This test case would pass if the backward inference setting also
restricted inference over relations to be backward only.