New way to specify a constraint's OntologySolver.
Now, as an alternative to the SOVLER_NAME::label naming convention,
it is possible to specify the solver for a constraint theough
an attribute and name the constraint anything.
Additionally, a mechanism is included for automatically setting
the solver name in cases where there is only a single OntologySolver
in the model.