Summary: Handle output ports that are spontaneous, thus fixing zero delay loop warnings.
See https://www.terraswarm.org/accessors/wiki/Main/CapeCodeZeroDelayLoops

The solution was to add declareDelayDependency() to the JavaScript
actor. First, we loop through the output ports and add a
SingletonParameter to the port if the spontaneous argument is present
in the accessor. Then, declareDelayDependency() loops through the
ports and declares any ports with spontaneous == true as not depending
on any of the inputs.